First public version of Note Registry
This is the first version that is available to everyone that wants to use it or give it a try. 
This is a small program that allows the user to write in a notepad style program with the ability to add the notes used to a recent notes section, for easy switching and easy writing.

The program is a project i am working on in my free time from school and was thought of, when my teacher mentioned writing notes to help remember what was taught in the various classes + the various commands from different systems in for example cisco servers.

There is a variety of ways to use the program. Everything in the program is meant to be easy to use and allow the user to work efficiently.

------Shortcuts in program------
- New file: Ctrl + N
- Save file: Ctrl + S
- Save file as: Ctrl + Shift + S
- Open file: Ctrl + O
- Recent note (Add): Ctrl + Alt + S
- Recent note (Remove): Ctrl + Alt + D"

------Planned features------
- Text customization

Got any ideas or features you would like to see?
Get in contact ^^.